Plot result of PCA

This function is internally called by runDESeq2

plotPCA(rld, pheno_info, out_dir, prefix = "DESeq2")

Arguments

rld
Count table after 'regularized log transformation' by rlog

pheno_info
A data frame with sample identifiers as row names including maximal two columns defining (sub-)conditions. The sample identifiers must be identical to those in the rld table.

out_dir
Directory for PDF file including the PCA plot

prefix
Prefix for name of PDF file

Value

DESeq2_PCA.pdf in output directory