Processes CD-HIT result file

This function is internally called by runClustering

processClusters(map, clusters, out_dir)

Arguments

map
A data frame with sequences as row names and sequence identifiers in first column. Can be generated by createMap

out_dir
Directory of CD-HIT result file and where generated files will be saved

Value

A data frame with the columns 'qseqid', 'cl_id' and 'sequences' containing the sequence identifier, the sequence and the assigned cluster identifier.