Run BLAST for small RNA sequences

This function takes a vector of sequences in FASTA style as input and applies BLAST running on a local blast database. By default, the parameters are configured for blasting small RNA sequences of strand-specific RNA-seq experiments with 100% identity over the whole query sequence.

runBlast(blast_exec, blast_db, ncores, sig_sequences, strand = "plus",
  identity = 100, dust = "no", soft_masking = "false")

Arguments

blast_exec
Path to your installation of the 'blastn' binary

blast_db
Path to your local BLAST database

ncores
Number of cores used by BLAST

sig_sequences
Vector of sequences in FASTA style generated by the sequencesAsFasta

strand
Query strand orientation to search against database. By default 'plus', assuming underlying strand-specific RNA-seq experiments.

identity
Identity cutoff applied for filtering BLAST results. By default 100%, to find only exact matches over the whole length of the query sequence.

dust
The parameter specifies whether low complexity filtering is applied to the query sequence. By default 'no', to prevent false-negative annotations of sequences of interest.

soft_masking
The parameter specifies whether low complexity filter is applied to the database. By default 'no', to prevent false-negative annotations of sequences of interest.

Value

Each row of the resulting data frame corresponds to a BLAST hit including the columns 'qseqid', 'qlen', 'pident', 'length', 'mismatch', 'gapopen', 'qstart', 'qend', 'sstart', 'send', 'evalue' and 'bitscore'.