Function to execute CD-HIT Clustering

Sequentially, each sequency is either assinged to an existing cluster or is classified as a new cluster representative if no matching cluster can be found.

runClustering(cdhit_path, sequences, out_dir, identityCutoff, lengthCutoff,
  wordlength, map, optional = "")

Arguments

cdhit_path
Path to cd-hit-est executable

sequences
Vector of sequences in fasta style generated by the sequencesAsFasta

out_dir
Output folder

identityCutoff
Sequence identity cutoff used for clustering

lengthCutoff
Length difference cutoff

wordlength
CD-Hit word length

map
Data frame with sequences as row names and some identifier for each sequence in the first column

optional
Optional execution parameters

Examples

map <- createMap(sigResults)

#> Error in nrow(countTable): object 'sigResults' not found

sigSeqFasta <- sequencesAsFasta(sigResults,map)

#> Error in paste(">", map[row.names(sigResults), 1], sep = ""): object 'map' not found

cd_hit <- "$HOME/software/cdhit/cd-hit-est"
clustResult<-runClustering(cd_hit, sigSeqFasta, out_dir, 0.9, 0.9, 9, map)

#> Error in paste(out_dir, "sig_sequences.fa", sep = "/"): object 'out_dir' not found